+++ connect/amazon/xray/client/src/main/kotlin/org/http4k/connect/amazon/xray/model/TimeRangeType.kt
@@ -1,0 +1,5 @@
+package org.http4k.connect.amazon.xray.model
+
+enum class TimeRangeType {
+    TraceId, Event, Service
+}